> Updated the documentation as per the review comments. > Processing the Rest response to SlackResponse to handle failure causes from the slack apis.
